Split cluster support in a separate gem
