+++ Rakefile
@@ -1,10 +1,12 @@
 # frozen_string_literal: true
 
 require 'bundler/gem_tasks'
+Bundler::GemHelper.install_tasks(dir: "redis_cluster", name: "redis_cluster")
+
 require 'rake/testtask'
 
 namespace :test do
-  groups = %i(redis distributed sentinel cluster)
+  groups = %i(redis distributed sentinel)
   groups.each do |group|
     Rake::TestTask.new(group) do |t|
       t.libs << "test"
@@ -17,6 +19,13 @@ namespace :test do
   lost_tests = Dir["test/**/*_test.rb"] - groups.map { |g| Dir["test/#{g}/**/*_test.rb"] }.flatten
   unless lost_tests.empty?
     abort "The following test files are in no group:\n#{lost_tests.join("\n")}"
+  end
+
+  Rake::TestTask.new(:cluster) do |t|
+    t.libs << "redis_cluster/test" << "test"
+    t.libs << "redis_cluster/lib" << "lib"
+    t.test_files = FileList["redis_cluster/test/**/*_test.rb"]
+    t.options = '-v' if ENV['CI'] || ENV['VERBOSE']
   end
 end
