+++ redis_cluster/Gemfile
@@ -8,6 +8,4 @@ gem 'minitest'
 gem 'rake'
 gem 'rubocop', '~> 1.25.1'
 gem 'mocha'
-
-gem 'redis-client', github: 'redis-rb/redis-client'
-gem 'hiredis-client'
+gem 'redis-cluster-client', github: 'redis-rb/redis-cluster-client'
