+++ redis_cluster/lib/redis/cluster/version.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require "redis/version"
+
+class Redis
+  class Cluster
+    VERSION = Redis::VERSION
+  end
+end
