+++ redis_cluster/lib/redis_cluster.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require "redis/cluster"
