+++ redis_cluster/test/abnormal_state_test.rb
@@ -51,7 +51,7 @@ class TestClusterAbnormalState < Minitest::Test
 
   def test_raising_error_when_nodes_are_not_cluster_mode
     assert_raises(Redis::Cluster::InitialSetupError) do
-      build_another_client(cluster: %W[redis://127.0.0.1:#{PORT}])
+      build_another_client(nodes: %W[redis://127.0.0.1:#{PORT}])
     end
   end
 end
