+++ redis_cluster/test/blocking_commands_test.rb
