+++ redis_cluster/test/client_pipelining_test.rb
