+++ redis_cluster/test/client_replicas_test.rb
