+++ redis_cluster/test/client_transactions_test.rb
