+++ redis_cluster/test/commands_on_cluster_test.rb
