+++ redis_cluster/test/commands_on_connection_test.rb
