+++ redis_cluster/test/commands_on_geo_test.rb
