+++ redis_cluster/test/commands_on_hashes_test.rb
