+++ redis_cluster/test/commands_on_hyper_log_log_test.rb
