+++ redis_cluster/test/commands_on_keys_test.rb
