+++ redis_cluster/test/commands_on_lists_test.rb
