+++ redis_cluster/test/commands_on_scripting_test.rb
