+++ redis_cluster/test/commands_on_server_test.rb
