+++ redis_cluster/test/commands_on_sets_test.rb
