+++ redis_cluster/test/commands_on_sorted_sets_test.rb
