+++ redis_cluster/test/commands_on_streams_test.rb
