+++ redis_cluster/test/commands_on_strings_test.rb
