+++ redis_cluster/test/commands_on_transactions_test.rb
