+++ redis_cluster/test/commands_on_value_types_test.rb
